- Main view - details panel implementation(http://jira.skobbler.de:8080/browse/JBOSCH-18) - Button panel - GUI implementation (http://jira.skobbler.de:8080/browse/JBOSCH-24) - Road sign search method + GUI implementation (http://jira.skobbler.de:8080/browse/JBOSCH-15)
git-svn-id: svn://svn-srv/skosigns/trunk@8 702c9929-26ad-4ec3-b8b7-334534ed37ec

+++ src/org/openstreetmap/josm/plugins/skosigns/entity/RoadSign.java
@@ -48,7 +48,7 @@ public class RoadSign {
     private String type;
     private Long tstamp;
     private Image image;
-    private Byte confidence;
+    private Short confidence;
     private CarPosition carPos;
     private SignPosition signPos;
     private Collection<LatLon> nearbyPos;
@@ -101,7 +101,7 @@ public class RoadSign {
      * comments posted by the users related to the road sign
      */
     public RoadSign(Long id, String type, Long tstamp, Image image,
-            Byte confidence, CarPosition carPos, SignPosition signPos,
+            Short confidence, CarPosition carPos, SignPosition signPos,
             Collection<LatLon> nearbyPos, Trip trip, Status status,
             Long duplicateOf, Collection<Comment> comments) {
         this(id, type, signPos, status);
@@ -134,7 +134,7 @@ public class RoadSign {
         return image;
     }
     
-    public Byte getConfidence() {
+    public Short getConfidence() {
         return confidence;
     }
